+++ internal/test/benchmark/benchmark_storage_test.go
@@ -6,21 +6,18 @@ import (
 	"testing"
 	"time"
 
+	"cloud.google.com/go/bigquery/storage/managedwriter/adapt"
 	"github.com/OTA-Insight/bqwriter"
-	"google.golang.org/protobuf/reflect/protodesc"
 )
 
 // NOTE: created an issue about nested type errors:
 // https://github.com/googleapis/google-cloud-go/issues/5097
 
 func BenchmarkStorageStreamerDefault(b *testing.B) {
-	// TODO: how can we have these nested types automatically included?
-	protoDescriptor := protodesc.ToDescriptorProto((&TemporaryDataProto3{}).ProtoReflect().Descriptor())
-	// protoDescriptor.NestedType = append(
-	// 	protoDescriptor.NestedType,
-	// 	protodesc.ToDescriptorProto((&timestamppb.Timestamp{}).ProtoReflect().Descriptor()),
-	// 	protodesc.ToDescriptorProto((&TemporaryDataParameterProto3{}).ProtoReflect().Descriptor()),
-	// )
+	protoDescriptor, err := adapt.NormalizeDescriptor((&TemporaryDataProto2{}).ProtoReflect().Descriptor())
+	if err != nil {
+		b.Fatalf("failed to create normalized descriptor: %v", err)
+	}
 	streamer, err := bqwriter.NewStreamer(
 		context.Background(),
 		benchmarkBigQueryProjectID,
@@ -59,13 +56,10 @@ func BenchmarkStorageStreamerDefaultJson(b *testing.B) {
 }
 
 func benchmarkStorageStreamerForParameters(b *testing.B, workerCount int, workerQueueSize int, maxBatchDelay time.Duration) {
-	// TODO: how can we have these nested types automatically included?
-	protoDescriptor := protodesc.ToDescriptorProto((&TemporaryDataProto3{}).ProtoReflect().Descriptor())
-	// protoDescriptor.NestedType = append(
-	// 	protoDescriptor.NestedType,
-	// 	protodesc.ToDescriptorProto((&timestamppb.Timestamp{}).ProtoReflect().Descriptor()),
-	// 	protodesc.ToDescriptorProto((&TemporaryDataParameterProto3{}).ProtoReflect().Descriptor()),
-	// )
+	protoDescriptor, err := adapt.NormalizeDescriptor((&TemporaryDataProto2{}).ProtoReflect().Descriptor())
+	if err != nil {
+		b.Fatalf("failed to create normalized descriptor: %v", err)
+	}
 	streamer, err := bqwriter.NewStreamer(
 		context.Background(),
 		benchmarkBigQueryProjectID,